refactor: move accounttier to common
